Refactor: merge .env file of sampleGenerator into .env.test file
