+++ src/sampleGenerator/README.md
@@ -8,18 +8,16 @@
 
 **SETUP**
 
-1. 현재 디렉터리(README.md 파일이 있는 곳)에 .env 파일을 아래와 같이 추가합니다.
+1. *(optional)* Root directory의 `.env.test` 파일에 다음 내용을 추가합니다.
    ```
-   #mongoDB 경로 (직접 입력해야 함 - 기본값은 mongodb://localhost:27017/local)
-   DB_PATH=mongodb://localhost:27017/local
    #방과 각각의 방의 채팅 개수
-   NUM_OF_ROOMS=2
-   NUM_OF_CHATS=200
+   SAMPLE_NUM_OF_ROOMS=2
+   SAMPLE_NUM_OF_CHATS=200
    #채팅 간 최대 시간 간격(단위: 초, 소수도 가능)
-   MAXIMUM_INTERVAL_BETWEEN_CHATS=20
+   SAMPLE_MAXIMUM_INTERVAL_BETWEEN_CHATS=20
    #새로운 채팅이 각각 입/퇴장 메시지일 확률(각각 10%)
-   OCCURENCE_OF_JOIN=0.1
-   OCCURENCE_OF_ABORT=0.1
+   SAMPLE_OCCURENCE_OF_JOIN=0.1
+   SAMPLE_OCCURENCE_OF_ABORT=0.1
    ```
 1. sampleData.json에 장소, 유저, 방 데이터를 입력합니다.
    javascript `User { "id": "sampleId", 사용자 id }`
